ページ遷移後に天気予報を表示しCLOSEボタンを実装
